Fix some code smells

+++ src/bbs_util.h
@@ -206,13 +206,13 @@ typedef void (bbs_bn_prf)(bbs_cipher_suite_t *cipher_suite,
 			 );
 
 // Big endian conversion
-#define UINT64_H2BE(x) (((x & 0xff00000000000000LL) >> 56) | \
-			((x & 0x00ff000000000000LL) >> 40) | \
-			((x & 0x0000ff0000000000LL) >> 24) | \
-			((x & 0x000000ff00000000LL) >>  8) | \
-			((x & 0x00000000ff000000LL) <<  8) | \
-			((x & 0x0000000000ff0000LL) << 24) | \
-			((x & 0x000000000000ff00LL) << 40) | \
-			((x & 0x00000000000000ffLL) << 56))
+#define UINT64_H2BE(x) (((x & (uint64_t)0xff00000000000000LL) >> 56) | \
+			((x & (uint64_t)0x00ff000000000000LL) >> 40) | \
+			((x & (uint64_t)0x0000ff0000000000LL) >> 24) | \
+			((x & (uint64_t)0x000000ff00000000LL) >>  8) | \
+			((x & (uint64_t)0x00000000ff000000LL) <<  8) | \
+			((x & (uint64_t)0x0000000000ff0000LL) << 24) | \
+			((x & (uint64_t)0x000000000000ff00LL) << 40) | \
+			((x & (uint64_t)0x00000000000000ffLL) << 56))
 
 #endif /*BBS_UTIL_H*/
